--- mithril-stm/src/bls_multi_sig/mod.rs
@@ -1,5 +1,0 @@
-mod signing_key;
-mod verification_key;
-mod pop;
-mod verification_key_pop;
-mod signature;
