+++ mithril-stm/src/bls_multi_signature/signing_key.rs
@@ -1,17 +1,12 @@
-
+use crate::bls_multi_signature::signature::Signature;
 use crate::error::{blst_err_to_mithril, MultiSignatureError};
-use blst::min_sig::{
-    AggregatePublicKey, AggregateSignature, PublicKey as BlstVk, SecretKey as BlstSk,
-    Signature as BlstSig,
-};
-use blst::{blst_p1, blst_p2, p1_affines, p2_affines, BLST_ERROR};
+use blst::min_sig::SecretKey as BlstSk;
 use rand_core::{CryptoRng, RngCore};
-use crate::bls_multi_sig::signature::Signature;
 
 /// MultiSig secret key, which is a wrapper over the BlstSk type from the blst
 /// library.
 #[derive(Debug, Clone)]
-pub struct SigningKey(BlstSk);
+pub struct SigningKey(pub BlstSk);
 
 impl SigningKey {
     /// Generate a secret key
@@ -47,6 +42,6 @@ impl SigningKey {
     }
 
     pub(crate) fn to_blst_sk(&self) -> BlstSk {
-        self.0
+        self.0.clone()
     }
 }
