+++ mithril-stm/src/lib.rs
@@ -16,9 +16,7 @@ pub use crate::error::{
 };
 
 #[cfg(feature = "benchmark-internals")]
-pub mod multi_sig;
+pub mod bls_multi_signature;
 
 #[cfg(not(feature = "benchmark-internals"))]
-mod multi_sig;
-
-mod bls_multi_sig;
+mod bls_multi_signature;
